Started rebuilding ToastController in es2015 syntax, using webpack
