+++ src/ToastController.js
@@ -1,22 +1,25 @@
-(function(){
-    /**
-     * Creates a new ToastController, which is a manager for Toasts. Should only
-     * ever be one of these on a page.
-     * @class
-     */
-    var ToastController = function(){
+import {Toast} from "./Toast.js";
+
+/**
+ * Creates a new ToastController, which is a manager for Toasts. Should only
+ * ever be one of these on a page.
+ * @class
+ */
+export default class ToastController {
+
+    constructor() {
         this.activeToasts = {};
         this.iconBase = "material-icons";
 
         // Setup the container for Toasts
         this.createContainer();
-    };
+    }
 
     /**
      * Creates a hidden div on the page to contain Toasts inside of it. The div
      * will be a child of the document body.
      */
-    ToastController.prototype.createContainer = function() {
+    createContainer() {
         if(this.container) {
             return;
         }
@@ -25,7 +28,7 @@
         this.container = document.createElement("div");
         this.container.id = "__toast_container";
         body.appendChild(this.container);
-    };
+    }
 
     /**
      * Opens a new Toast based on the the configuration data passed in as a map
@@ -50,7 +53,7 @@
      *
      * @returns {Toast} The Toast that was created and opened
      */
-    ToastController.prototype.openToast = function(config){
+    openToast(config) {
         var toast = new Toast(config.name, config.text, this);
 
         if(config.icon) {
@@ -74,18 +77,18 @@
         this.activeToasts[name] = toast;
 
         return toast;
-    };
+    }
 
     /**
      * Closes an existing Toast given the Toast's name (assuming it exists)
      *
      * @param {string} name the name of the Toast set when calling openToast
      */
-    ToastController.prototype.closeToast = function(name){
+    closeToast(name) {
         if(this.activeToasts[name]){
             this.activeToasts[name].close();
         }
-    };
+    }
 
     /**
      * Alias to open a "success" type Toast, which has a low priority, green
@@ -95,14 +98,14 @@
      *
      * @returns {Toast}
      */
-    ToastController.prototype.success = function(text) {
+    success(text) {
         return this.openToast({
             name: "sucess",
             text: text,
             icon: "done",
             className: "success"
         });
-    };
+    }
 
     /**
      * Alias to open an "error" type Toast, which has a high priority, red
@@ -112,7 +115,7 @@
      *
      * @returns {Toast}
      */
-    ToastController.prototype.error = function(text) {
+    error(text) {
         return this.openToast({
             name: "error",
             text: text,
@@ -120,7 +123,7 @@
             className: "error",
             priority: window.ToastPriority.HIGH
         });
-    };
+    }
 
     /**
      * Alias to open a "warning" type Toast, which has a high priority, yellow
@@ -130,7 +133,7 @@
      *
      * @returns {Toast}
      */
-    ToastController.prototype.warn = function(text) {
+    warn(text) {
         return this.openToast({
             name: "warning",
             text: text,
@@ -138,176 +141,6 @@
             className: "warn",
             priority: window.ToastPriority.HIGH
         });
-    };
-
-    /**
-     * Creates a new Toast object based on the configuration passed in.
-     *
-     * @class
-     *
-     * @param {string} name the name of the Toast (used programmatically, never
-     * displayed)
-     *
-     * @param {string} text the text to appear inside the Toast
-     *
-     * @param {ToastController} controller the controller that manages this
-     * Toast (so the Toast can tell its controller that it is going to destroy
-     * itself)
-     */
-    var Toast = function(name, text, controller){
-        this.className = "";
-        this.closed = false;
-        this.controller = controller;
-        this.delay = 5000;
-        this.icon = false;
-        this.isOpen = false;
-        this.name = name;
-        this.text = text;
-        this.priority = window.ToastPriority.LOW;
-    };
-
-    /**
-     * Actually spawns the Toast's HTMLElement in the DOM, and starts the timer
-     * for it to dissapear.
-     */
-    Toast.prototype.create = function(){
-        var self = this;
-        var textContainer = document.createElement("span");
-        this.element = document.createElement("div");
-
-        this.assignPriority();
-
-        this.element.className = "toast " + this.className;
-        textContainer.className = "text";
-        textContainer.innerHTML = this.text;
-
-        if(this.icon) {
-            this.element.appendChild(this.buildIcon());
-        }
-
-        this.element.appendChild(textContainer);
-
-        // Bindings
-        this.element.addEventListener("transitionend", function(e){
-            self.transitionEnd(e);
-        }, false);
-
-        this.element.onclick = function(e){
-            self.onClick(e);
-        };
-
-        // Finally, open it up
-        requestAnimationFrame(function(){
-            self.controller.container.appendChild(self.element);
-
-            self.open();
-        });
-
-        setTimeout(function(){
-            self.close();
-        }, this.delay);
-    };
-
-    /**
-     * Sets the primary element for the Toast to have the proper priority and
-     * ARIA attributes.
-     */
-    Toast.prototype.assignPriority = function() {
-        if(this.priority === window.ToastPriority.HIGH) {
-            this.element.setAttribute("role", "alert");
-        } else {
-            this.element.setAttribute("role", "dialog");
-            this.element.setAttribute("aria-label", "Toast Notification");
-        }
-    };
-
-    /**
-     * Builds the <i> icon HTMLElement for the icon specified for the Toast
-     *
-     * @returns {HTMLElement} the "i" tag element that represents the icon
-     */
-    Toast.prototype.buildIcon = function() {
-        var icon = document.createElement("i");
-
-        icon.className = window.ToastController.iconBase + " icon";
-        icon.innerHTML = this.icon;
-
-        icon.setAttribute("aria-hidden", "true");
-
-        return icon;
-    };
-
-    /**
-     * Handler for "onclick" events fired on the Toast; will close the Toast
-     * when the Toast is clicked.
-     */
-    Toast.prototype.onClick = function(e){
-        this.close();
-    };
-
-    /**
-     * Handler for "transitionend" events fired on the Toast; will destroy the
-     * Toast assuming that the Toast just closed and finished its closing
-     * animation.
-     */
-    Toast.prototype.transitionEnd = function(e){
-        if(!this.isOpen){
-            // destroy element and tell controller to forget about the Toast
-            this.element.parentElement.removeChild(this.element);
-
-            this.controller.activeToasts[this.name] = undefined;
-        }
-    };
-
-    /**
-     * Opens the Toast (assumes that the Toast's DOM compontent has been
-     * generated).
-     */
-    Toast.prototype.open = function(){
-        var self = this;
-
-        this.isOpen = true;
-
-        requestAnimationFrame(function(){
-            self.element.className = "toast open " + self.className;
-        });
-    };
-
-    /**
-     * Closes the Toast. Warning: once the Toast is finished closing, it will
-     * destroy itself.
-     */
-    Toast.prototype.close = function(){
-        if(this.closed){
-            return;
-        }
-
-        var self = this;
-
-        this.isOpen = false;
-        this.closed = true;
-
-        requestAnimationFrame(function(){
-            self.element.className = "toast " + self.className;
-        });
-    };
-
-    function init() {
-        if(!window.ToastController){
-            window.ToastController = new ToastController();
-
-            window.ToastPriority = {
-                HIGH: 1,
-                LOW: 0
-            };
-        }
     }
 
-    if(document.readyState === "loading") {
-        document.addEventListener("DOMContentLoaded", function() {
-            init();
-        }, false);
-    } else {
-        init();
-    }
-})();
+}
